frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.5"
dependencies = [
    "markdown>=3.10",
    "plotly>=6.2.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests-toolbelt>=1.0.0",
    "ruff>=0.12.8",
    "st-copy>=1.1.2",
    "streamlit==1.51.0",
    "streamlit-js-eval>=0.1.7",
]

[tool.ruff]
line-length = 120
target-version = "py312"
lint.select = ["ALL"]
